Add Instance Segmentation workflow wrapper on Object Detection
Adds an experimental Instance Segmentation workflow which delegates to
prebuilt Object Detection workflow. The Object Detection prebulit is
capable of handling both bounding boxes and polygons, however it is
unintuitive that a Object Detection workflow be used for polygons.
This commit adds a thin wrapper on Object Detection to properly
classify the workflow as "Instance Segmentation".