Add optional config flag to choose evaluator for Object Detection (#353)
Adds an optional boolean to the evaluator configuration used in the prebuilt Object Detection workflow.
The flag can be used to explicitly determine if the SingleClass or
Multiclass evaluator is run. In the absence of this flag, the
evaluator used is determined dynamically based on the number of
distinct classification labels found in the ground truths and inferences.