Update `kolena.initialize` verbose default to True
Updates `kolena.initialize` verbose default value to True. The
majority of calls to `kolena.initialize` in this package and docs use
`verbose=True` since it enables logging to stdout/stderr providing
valuable feedback to the user. This updates `kolena.initialize` to
default to using `verbose=True` to match the most common invocation pattern.