Fix semantic segmentation dataset example breakage
Fixes semantic segmentation dataset example break caused by
https://github.com/kolenaIO/kolena/pull/546. This commit prioritizes
inferring columns beginning or ending with `id` as the id_field over `locator`. This caused a
mismatch between the id field in the dataset and results for this
example.

Fixed by explicitly setting the id_field to `locator` when uploading
the dataset.